remove unnecassary boxing

+++ code/src/java/pcgen/gui2/tabs/SkillInfoTab.java
@@ -605,7 +605,7 @@ public class SkillInfoTab extends FlippingSplitPane implements CharacterInfoTab,
 			SkillCost cost = levels.getSkillCost(targetLevel, skill);
 			if (value < 0)
 			{
-				value = Float.valueOf(0);
+				value = 0f;
 			}
 			float max =
 					levels.getMaxRanks(targetLevel, cost,
